requirements.txt
asgiref==3.11.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
distlib==0.4.0
Django==6.0
django-cors-headers==4.3.1
django-filter==25.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
filelock==3.20.0
inflection==0.5.1
openpyxl==3.1.5
oracledb==3.4.1
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pycparser==2.23
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.7
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.35.3
gunicorn==23.0.0
whitenoise==6.8.2
